Fix case mismatch
src/captureofficers/config/FactionConfigLoader.kt tries to load a JSON
file for each entry in the list in modSettings.json, but the case of the
strings in modSettings.json doesn't match the case of the files in the
config directory, which means some files won't be found on case
sensitive filesystems.

This commit makes all the factionIds in modSettings.json lowercase. I
didn't change the loading behaviour when FactionConfig is instantiated
with `loadFromJson=true` because that never occurs within this repo, so
I'm assuming it's a public interface. Also, I don't know how the "path"
argument to getMergedJSONForMod() is interpreted.

If paths to getMergedJSONForMod must match case-sensitively with files
in mod directories (on case-sensitive filesystems) and
`FactionConfig(my_faction_id, true)` is something that gets called by
third parties then it might be better to change the filenames of the
pre-packaged files to exactly match the faction ids that each faction
uses.

Reported by: https://fractalsoftworks.com/forum/index.php?topic=25979.msg391519#msg391519